Farm Calculator (In development)
Goal: Create a GUI that workers could use to enter gallons and quarts of strawberries picked. Allow the user to select the quality of the strawberries and the application to set the price based on quality. Allow the application to report amount earned to the individual that picked the strawberries.